+++ src/dhad/assets/starter/js/dhad.runtime.js
@@ -1545,7 +1545,7 @@
   }
 
   const api = Object.freeze({
-    version: "3.0.0",
+    version: "3.0.1",
     setTheme,
     toggleTheme,
     configureSymbols,